modified the process of managing the widgets.
Now the widget can be added to the system, and registered in the system
by an API : system.logWidget(). In my design, widgets can be made by
anyone, but they don't need to know the details of the system. only
define how to open, how to close and so on.